+++ Modules/Seo/Http/Controllers/Admin/SeoRuleController.php
@@ -8,18 +8,20 @@ use Modules\Seo\Http\Requests\Admin\SeoRuleCreateRequest;
 use Modules\Seo\Http\Requests\Admin\SeoRuleUpdateRequest;
 use Modules\Seo\Http\Resources\SeoRuleResource;
 use Modules\Seo\Models\SeoRule;
+use Modules\Seo\Repositories\SeoRuleRepository;
 use Modules\Seo\Services\SeoRuleStorage;
 
 class SeoRuleController extends Controller
 {
     public function __construct(
-        protected SeoRuleStorage $seoRuleStorage
-    ) {
-        $this->authorizeResource(SeoRule::class, 'seo_rule');
-    }
+        protected SeoRuleStorage $seoRuleStorage,
+        protected SeoRuleRepository $seoRuleRepository
+    ) {}
 
     public function store(SeoRuleCreateRequest $request)
     {
+        $this->authorize('create', SeoRule::class);
+
         $seoRuleDto = SeoRuleDto::fromFormRequest($request);
 
         if (!$seoRuleDto->assigned_by_id) {
@@ -31,15 +33,23 @@ class SeoRuleController extends Controller
         return new SeoRuleResource($seoRule);
     }
 
-    public function update(SeoRule $seo_rule, SeoRuleUpdateRequest $request)
+    public function update(int $seo_rule, SeoRuleUpdateRequest $request)
     {
+        $seoRule = $this->seoRuleRepository->find($seo_rule);
+
+        $this->authorize('update', $seoRule);
+
         $seo_rule = $this->seoRuleStorage->update($seo_rule, (new SeoRuleDto($request->validated()))->only(...$request->keys()));
 
         return new SeoRuleResource($seo_rule);
     }
 
-    public function destroy(SeoRule $seo_rule)
+    public function destroy(int $seo_rule)
     {
+        $seoRule = $this->seoRuleRepository->find($seo_rule);
+
+        $this->authorize('delete', $seoRule);
+
         $this->seoRuleStorage->delete($seo_rule);
 
         return response()->noContent();
